// backend/src/schemas/authSchema.ts
import z from "zod";


export const signupSchema = z.object({
    name: z.string().min(2, "Name must be at least 2 characters"),
    email: z.email("Invalid email format"),
    password: z.string().min(6, "Password must be at least 6 characters"),
});
  
export const loginSchema = z.object({
    email: z.email("Invalid email format"),
    password: z.string("Password is required"),
});